refactor: consolidate rectangular meshes — kernel-CDF takes the plain names
RectangularAdaptDensity/AdaptImage are now the kernel-density-CDF meshes
(formerly RectangularKernelAdapt*, #374); RectangularUniform is unchanged.
Removed: the linear empirical-CDF implementation, the spline-CDF pair
(#289), the PCA-rotated multi-node mesh (#323), density_components, and
their interpolators/geometry/tests. InterpolatorRectangular keeps its
name/module path as the kernel implementation; prior-config keys and
downstream imports are unchanged.

Closes #402.
