+++ autoarray/inversion/mesh/interpolator/rectangular.py
@@ -1,19 +1,43 @@
+"""
+The adaptive rectangular interpolator, built on a kernel-density CDF transform.
+
+The adaptive rectangular mesh transforms source-plane coordinates through a
+per-axis CDF so that the uniform mesh pixels adapt to the density (or weight)
+of the traced points.  The CDF is a smooth kernel-density CDF (the RTU
+formulation of Enzi et al., arXiv:2606.30620): per axis
+
+    F(x) = sum_i w_i * Phi((x - x_i) / h)
+
+where ``Phi`` is the standard normal CDF, ``x_i`` the traced points, ``w_i``
+uniform weights (density adaption) or the normalized adapt-image weights
+(image adaption), and ``h`` a bandwidth tied to the mesh resolution.
+
+Properties, by construction:
+
+- strictly monotone;
+- C-infinity in the queries AND the traced-point positions — no ranks, no
+  sorts, no ``argsort`` anywhere, so there is nothing to swap and the
+  likelihood carries smooth mass/shear gradients in every configuration
+  (including pixelization over-sampling 1 and the interferometer sparse
+  path, where an empirical point-rank CDF is exactly piecewise-constant);
+- duplicate-safe: coincident traced points simply stack their weights.
+
+The forward transform is evaluated exactly at the query points (keeping the
+C-infinity guarantee) and rescaled so the data bounding box maps onto the
+unit square exactly, clamping to [0, 1] outside the data range.  The
+inverse — only needed at fixed unit-square grid values (mesh pixel
+centres/edges) — is a linear-interp lookup on a small fixed table of
+``n_knots`` knots spanning the data range; because the inverse queries are
+constants, gradients flow smoothly through the table values.
+"""
+
 import numpy as np
 from functools import partial
+from typing import Optional
 
 from autonerves import cached_property
 
 from autoarray.inversion.mesh.interpolator.abstract import AbstractInterpolator
-
-
-def forward_interp(xp, yp, x):
-
-    import jax
-    import jax.numpy as jnp
-
-    return jax.vmap(jnp.interp, in_axes=(1, 1, 1, None, None), out_axes=(1))(
-        x, xp, yp, 0, 1
-    )
 
 
 def reverse_interp(xp, yp, x):
@@ -21,26 +45,6 @@ def reverse_interp(xp, yp, x):
     import jax.numpy as jnp
 
     return jax.vmap(jnp.interp, in_axes=(1, 1, 1), out_axes=(1))(x, xp, yp)
-
-
-def forward_interp_np(xp, yp, x):
-    """
-    xp: (N, M)
-    yp: (N, M)
-    x : (M,)  ← one x per column
-    """
-
-    if yp.ndim == 1 and xp.ndim == 2:
-        yp = np.broadcast_to(yp[:, None], xp.shape)
-
-    K, M = x.shape
-
-    out = np.empty((K, 2), dtype=xp.dtype)
-
-    for j in range(2):
-        out[:, j] = np.interp(x[:, j], xp[:, j], yp[:, j], left=0, right=1)
-
-    return out
 
 
 def reverse_interp_np(xp, yp, x):
@@ -67,41 +71,151 @@ def reverse_interp_np(xp, yp, x):
     return out
 
 
-def create_transforms(traced_points, mesh_weight_map=None, xp=np):
+KERNEL_CDF_DEFAULT_BANDWIDTH: float = 1.0
+KERNEL_CDF_DEFAULT_KNOTS: int = 64
 
-    N = traced_points.shape[0]  # // 2
+# Queries per block of the forward-transform evaluation. The exact kernel sum
+# broadcasts an (M, N, 2) array; blocking the query dimension caps the peak at
+# (block, N, 2) — 512 queries × 15.4k points × 2 axes × 8 B ≈ 126 MB at
+# production imaging scale (previously ~60 GB unblocked at M ≈ 246k). Values
+# are identical to float precision: the sum over points is unchanged, blocks
+# only tile the query axis.
+KERNEL_FORWARD_BLOCK: int = 512
+
+_SQRT2 = np.sqrt(2.0)
+
+
+def _norm_cdf(t, xp):
+    """Standard normal CDF, xp-aware (scipy erf on numpy, jax.scipy under jax)."""
+    if xp is np:
+        from scipy.special import erf
+    else:
+        from jax.scipy.special import erf
+
+    return 0.5 * (1.0 + erf(t / _SQRT2))
+
+
+def create_transforms(
+    traced_points,
+    mesh_pixels: int,
+    mesh_weight_map=None,
+    bandwidth: float = KERNEL_CDF_DEFAULT_BANDWIDTH,
+    n_knots: int = KERNEL_CDF_DEFAULT_KNOTS,
+    xp=np,
+):
+    """
+    Build the per-axis kernel-density CDF transform pair.
+
+    The returned ``transform`` maps (scaled) source-plane coordinates into the
+    unit square and ``inv_transform`` maps unit-square coordinates back to the
+    (scaled) source plane.
+
+    Parameters
+    ----------
+    traced_points
+        The (N, 2) scaled source-plane coordinates the CDF adapts to.
+    mesh_pixels
+        Mesh pixels per axis; sets the default bandwidth scale
+        ``h = bandwidth * data_span / mesh_pixels`` per axis.
+    mesh_weight_map
+        Optional (N,) weights from the adapt image (image adaption). ``None``
+        gives uniform weights (density adaption).
+    bandwidth
+        Bandwidth in units of the mesh pixel scale (data span / mesh_pixels).
+        Smaller values track the point density more sharply (approaching the
+        empirical CDF and its staircase); larger values smooth the mesh
+        geometry towards uniform.
+    n_knots
+        Size of the fixed knot table used to invert the CDF.
+    xp
+        The array library to use (numpy or jax.numpy).
+    """
+    points = traced_points
+    N = points.shape[0]
 
     if mesh_weight_map is None:
-        t = xp.arange(1, N + 1) / (N + 1)
-        t = xp.stack([t, t], axis=1)
-        sort_points = xp.sort(traced_points, axis=0)  # [::2]
+        w = xp.full((N,), 1.0 / N)
     else:
-        sdx = xp.argsort(traced_points, axis=0)
-        sort_points = xp.take_along_axis(traced_points, sdx, axis=0)
-        t = xp.stack([mesh_weight_map, mesh_weight_map], axis=1)
-        t = xp.take_along_axis(t, sdx, axis=0)
-        t = xp.cumsum(t, axis=0)
+        w = mesh_weight_map / xp.sum(mesh_weight_map)
+
+    lo = xp.min(points, axis=0)
+    hi = xp.max(points, axis=0)
+    span = hi - lo
+    h = bandwidth * span / mesh_pixels
+
+    def F_raw_block(q_block):
+        t = (q_block[:, None, :] - points[None, :, :]) / h[None, None, :]
+        return xp.sum(w[None, :, None] * _norm_cdf(t, xp), axis=1)
 
     if xp.__name__.startswith("jax"):
-        transform = partial(forward_interp, sort_points, t)
-        inv_transform = partial(reverse_interp, t, sort_points)
+
+        def F_raw(q):
+            import jax
+
+            M = q.shape[0]
+            n_blocks = -(-M // KERNEL_FORWARD_BLOCK)
+            pad = n_blocks * KERNEL_FORWARD_BLOCK - M
+            q_padded = xp.pad(q, ((0, pad), (0, 0)))
+            blocks = q_padded.reshape(n_blocks, KERNEL_FORWARD_BLOCK, 2)
+            out = jax.lax.map(F_raw_block, blocks)
+            return out.reshape(n_blocks * KERNEL_FORWARD_BLOCK, 2)[:M]
+
+    else:
+
+        def F_raw(q):
+            return np.concatenate(
+                [
+                    F_raw_block(q[i : i + KERNEL_FORWARD_BLOCK])
+                    for i in range(0, q.shape[0], KERNEL_FORWARD_BLOCK)
+                ],
+                axis=0,
+            )
+
+    # The unit square maps onto the data bounding box exactly (the kernel
+    # tails outside [lo, hi] are absorbed by the rescale).
+    u = xp.linspace(0.0, 1.0, n_knots)
+    knots = lo[None, :] + u[:, None] * span[None, :]
+
+    F_knots_raw = F_raw(knots)
+    F_lo = F_knots_raw[0]
+    F_hi = F_knots_raw[-1]
+    denom = F_hi - F_lo
+
+    F_knots = (F_knots_raw - F_lo[None, :]) / denom[None, :]
+
+    def transform(q):
+        F_q = (F_raw(q) - F_lo[None, :]) / denom[None, :]
+        return xp.clip(F_q, 0.0, 1.0)
+
+    if xp.__name__.startswith("jax"):
+        inv_transform = partial(reverse_interp, F_knots, knots)
         return transform, inv_transform
 
-    transform = partial(forward_interp_np, sort_points, t)
-    inv_transform = partial(reverse_interp_np, t, sort_points)
+    inv_transform = partial(reverse_interp_np, F_knots, knots)
     return transform, inv_transform
 
 
 def adaptive_rectangular_transformed_grid_from(
-    data_grid, grid, mesh_weight_map=None, xp=np
+    data_grid,
+    grid,
+    mesh_pixels: int,
+    mesh_weight_map=None,
+    bandwidth: float = KERNEL_CDF_DEFAULT_BANDWIDTH,
+    n_knots: int = KERNEL_CDF_DEFAULT_KNOTS,
+    xp=np,
 ):
-
+    """Map unit-square coordinates back to the source plane via the kernel CDF."""
     mu = data_grid.mean(axis=0)
     scale = data_grid.std(axis=0).min()
     source_grid_scaled = (data_grid - mu) / scale
 
-    transform, inv_transform = create_transforms(
-        source_grid_scaled, mesh_weight_map=mesh_weight_map, xp=xp
+    _, inv_transform = create_transforms(
+        source_grid_scaled,
+        mesh_pixels=mesh_pixels,
+        mesh_weight_map=mesh_weight_map,
+        bandwidth=bandwidth,
+        n_knots=n_knots,
+        xp=xp,
     )
 
     def inv_full(U):
@@ -110,11 +224,50 @@ def adaptive_rectangular_transformed_grid_from(
     return inv_full(grid)
 
 
+def adaptive_rectangular_areas_from(
+    source_grid_shape,
+    data_grid,
+    mesh_weight_map=None,
+    bandwidth: float = KERNEL_CDF_DEFAULT_BANDWIDTH,
+    n_knots: int = KERNEL_CDF_DEFAULT_KNOTS,
+    xp=np,
+):
+    """The source-plane area of every mesh pixel under the kernel-CDF transform."""
+    edges_y = xp.linspace(1, 0, source_grid_shape[0] + 1)
+    edges_x = xp.linspace(0, 1, source_grid_shape[1] + 1)
+
+    mu = data_grid.mean(axis=0)
+    scale = data_grid.std(axis=0).min()
+    source_grid_scaled = (data_grid - mu) / scale
+
+    _, inv_transform = create_transforms(
+        source_grid_scaled,
+        mesh_pixels=source_grid_shape[0],
+        mesh_weight_map=mesh_weight_map,
+        bandwidth=bandwidth,
+        n_knots=n_knots,
+        xp=xp,
+    )
+
+    def inv_full(U):
+        return inv_transform(U) * scale + mu
+
+    pixel_edges = inv_full(xp.stack([edges_y, edges_x]).T)
+    pixel_lengths = xp.diff(pixel_edges, axis=0).squeeze()
+
+    dy = pixel_lengths[:, 0]
+    dx = pixel_lengths[:, 1]
+
+    return xp.abs(xp.outer(dy, dx).flatten())
+
+
 def adaptive_rectangular_mappings_weights_via_interpolation_from(
     source_grid_size: int,
     data_grid,
     data_grid_over_sampled,
     mesh_weight_map=None,
+    bandwidth: float = KERNEL_CDF_DEFAULT_BANDWIDTH,
+    n_knots: int = KERNEL_CDF_DEFAULT_KNOTS,
     xp=np,
 ):
     """
@@ -136,7 +289,8 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
     Steps performed:
       1. Normalize the source-plane grid by subtracting its mean and dividing by
          the minimum axis standard deviation (to balance scaling).
-      2. Construct forward/inverse transforms which map the grid into the unit square [0,1]^2.
+      2. Construct the kernel-CDF forward/inverse transforms which map the grid
+         into the unit square [0,1]^2.
       3. Transform the oversampled source-plane grid into [0,1]^2, then scale it
          to index space `[0, source_grid_size)`.
       4. Compute floor/ceil along x and y axes to find the enclosing rectangular cell.
@@ -158,8 +312,12 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
     data_grid_over_sampled : (N, 2) ndarray
         Oversampled source-plane coordinates to be interpolated onto the rectangular grid.
     mesh_weight_map
-        The weight map used to weight the creation of the rectangular mesh grid, which is used for the
-        `RectangularBrightness` mesh which adapts the size of its pixels to where the source is reconstructed.
+        The weight map used to weight the creation of the rectangular mesh grid, which
+        adapts the size of the mesh pixels to where the source is reconstructed.
+    bandwidth
+        Bandwidth in units of the mesh pixel scale (see ``create_transforms``).
+    n_knots
+        Size of the fixed knot table used to invert the CDF.
 
     Returns
     -------
@@ -170,15 +328,19 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
         The bilinear interpolation weights for each of the four neighboring pixels.
         Order: [w_bl, w_br, w_tl, w_tr].
     """
-
     # --- Step 1. Normalize grid ---
     mu = data_grid.mean(axis=0)
     scale = data_grid.std(axis=0).min()
     source_grid_scaled = (data_grid - mu) / scale
 
     # --- Step 2. Build transforms ---
-    transform, inv_transform = create_transforms(
-        source_grid_scaled, mesh_weight_map=mesh_weight_map, xp=xp
+    transform, _ = create_transforms(
+        source_grid_scaled,
+        mesh_pixels=source_grid_size,
+        mesh_weight_map=mesh_weight_map,
+        bandwidth=bandwidth,
+        n_knots=n_knots,
+        xp=xp,
     )
 
     # --- Step 3. Transform oversampled grid into index space ---
@@ -200,9 +362,7 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
 
     # --- Step 6. Flatten indices ---
     def flatten(idx, n):
-        row = n - idx[:, 0]
-        col = idx[:, 1]
-        return row * n + col
+        return (n - idx[:, 0]) * n + idx[:, 1]
 
     flat_tl = flatten(idx_tl, source_grid_size)
     flat_tr = flatten(idx_tr, source_grid_size)
@@ -217,7 +377,6 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
     t_row = (grid_over_index[:, 0] - ix_down) / (ix_up - ix_down + 1e-12)
     t_col = (grid_over_index[:, 1] - iy_down) / (iy_up - iy_down + 1e-12)
 
-    # Weights
     w_tl = (1 - t_row) * (1 - t_col)
     w_tr = (1 - t_row) * t_col
     w_bl = t_row * (1 - t_col)
@@ -228,6 +387,18 @@ def adaptive_rectangular_mappings_weights_via_interpolation_from(
 
 
 class InterpolatorRectangular(AbstractInterpolator):
+    """Adaptive rectangular interpolator (kernel-density CDF).
+
+    A grid of (y,x) coordinates which represent an adaptive rectangular
+    pixelization, ordered such that pixels begin from the top-row and go
+    rightwards and then downwards. The mesh pixels adapt to the density (or
+    adapt-image weights) of the traced points through the smooth kernel-CDF
+    transform of ``create_transforms``.
+
+    This class is used in conjunction with the `inversion/pixelizations`
+    package to create rectangular pixelizations and mappers that perform an
+    `Inversion`.
+    """
 
     def __init__(
         self,
@@ -235,36 +406,11 @@ class InterpolatorRectangular(AbstractInterpolator):
         mesh_grid,
         data_grid,
         mesh_weight_map,
-        adapt_data: np.ndarray = None,
+        adapt_data: Optional[np.ndarray] = None,
+        bandwidth: float = KERNEL_CDF_DEFAULT_BANDWIDTH,
+        n_knots: int = KERNEL_CDF_DEFAULT_KNOTS,
         xp=np,
     ):
-        """
-        A grid of (y,x) coordinates which represent a uniform rectangular pixelization.
-
-        A `InterpolatorRectangular` is ordered such pixels begin from the top-row and go rightwards and then downwards.
-        It is an ndarray of shape [total_pixels, 2], where the first dimension of the ndarray corresponds to the
-        pixelization's pixel index and second element whether it is a y or x arc-second coordinate.
-
-        For example:
-
-        - grid[3,0] = the y-coordinate of the 4th pixel in the rectangular pixelization.
-        - grid[6,1] = the x-coordinate of the 7th pixel in the rectangular pixelization.
-
-        This class is used in conjuction with the `inversion/pixelizations` package to create rectangular pixelizations
-        and mappers that perform an `Inversion`.
-
-        Parameters
-        ----------
-        values
-            The grid of (y,x) coordinates corresponding to the centres of each pixel in the rectangular pixelization.
-        shape_native
-            The 2D dimensions of the rectangular pixelization with shape (y_pixels, x_pixel).
-        pixel_scales
-            The (y,x) scaled units to pixel units conversion factors of every pixel. If this is input as a `float`,
-            it is converted to a (float, float) structure.
-        origin
-            The (y,x) origin of the pixelization.
-        """
         super().__init__(
             mesh=mesh,
             mesh_grid=mesh_grid,
@@ -273,10 +419,11 @@ class InterpolatorRectangular(AbstractInterpolator):
             xp=xp,
         )
         self.mesh_weight_map = mesh_weight_map
+        self.bandwidth = bandwidth
+        self.n_knots = n_knots
 
     @cached_property
     def mesh_geometry(self):
-
         from autoarray.inversion.mesh.mesh_geometry.rectangular import (
             MeshGeometryRectangular,
         )
@@ -286,18 +433,21 @@ class InterpolatorRectangular(AbstractInterpolator):
             mesh_grid=self.mesh_grid,
             data_grid=self.data_grid,
             mesh_weight_map=self.mesh_weight_map,
+            kernel_bandwidth=self.bandwidth,
+            kernel_knots=self.n_knots,
             xp=self._xp,
         )
 
     @cached_property
     def _mappings_sizes_weights(self):
-
         mappings, weights = (
             adaptive_rectangular_mappings_weights_via_interpolation_from(
                 source_grid_size=self.mesh.shape[0],
                 data_grid=self.data_grid.array,
                 data_grid_over_sampled=self.data_grid.over_sampled.array,
                 mesh_weight_map=self.mesh_weight_map,
+                bandwidth=self.bandwidth,
+                n_knots=self.n_knots,
                 xp=self._xp,
             )
         )
