+++ autoarray/inversion/mesh/mesh_geometry/abstract.py
@@ -9,7 +9,6 @@ class AbstractMeshGeometry:
         mesh_grid,
         data_grid,
         mesh_weight_map=None,
-        spline_deg=None,
         kernel_bandwidth=None,
         kernel_knots=None,
         xp=np,
@@ -19,11 +18,8 @@ class AbstractMeshGeometry:
         self.mesh_grid = mesh_grid
         self.data_grid = data_grid
         self.mesh_weight_map = mesh_weight_map
-        # When non-None, rectangular geometry uses the spline-CDF helpers
-        # instead of the linear-interp CDF (areas / edges transforms only).
-        self.spline_deg = spline_deg
-        # When non-None, rectangular geometry uses the kernel-density-CDF
-        # helpers instead of the linear-interp CDF (areas / edges transforms).
+        # Kernel-density-CDF parameters for the rectangular geometry's
+        # areas / edges transforms; None falls back to the kernel defaults.
         self.kernel_bandwidth = kernel_bandwidth
         self.kernel_knots = kernel_knots
         self._use_jax = xp is not np
@@ -32,5 +28,6 @@ class AbstractMeshGeometry:
     def _xp(self):
         if self._use_jax:
             import jax.numpy as jnp
+
             return jnp
         return np
